SCMOD-11949: Add gosu to base image
Adding gosu to the base image. Gosu allows derived images to run commands as a specified user, rather than as the default user.

To use gosu, set the RUNAS_USER environment variable in the derived container's Dockerfile. Subsequent commands will then be run as the specified user:

ENV RUNAS_USER=my-user
CMD ["whoami"] # Outputs my-user

Note: the user specified by the RUNAS_USER is expected to already exist, and the CMD will fail if this is not the case.